feat(code_action): show as prompt
So that the user can search code action instead of navigating one by one.
